backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
gunicorn==23.0.0
django-cors-headers==4.4.0
Pillow==11.2.1
